--- server/src/main/java/net/hamsterapps/cedserver/exception/project/ProjectFoundException.java
@@ -1,37 +1,0 @@
-package net.hamsterapps.cedserver.exception.project;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import graphql.ErrorClassification;
-import graphql.ErrorType;
-import graphql.GraphQLError;
-import graphql.language.SourceLocation;
-
-public class ProjectFoundException extends RuntimeException implements GraphQLError {
-
-  private static final long serialVersionUID = 1L;
-
-  private Map<String, Object> extensions = new HashMap<>();
-
-  public ProjectFoundException(String message, Long projectId) {
-    super(message);
-    extensions.put("projectFound", projectId);
-  }
-
-  @Override
-  public List<SourceLocation> getLocations() {
-    return null;
-  }
-
-  @Override
-  public Map<String, Object> getExtensions() {
-    return extensions;
-  }
-
-  @Override
-  public ErrorClassification getErrorType() {
-    return ErrorType.DataFetchingException;
-  }
-}
